fix wrong type form

+++ models/user_shipping_location.go
@@ -6,8 +6,8 @@ type UserShippingLocation struct {
 	gorm.Model
 	PhoneNumber			string		`form:"phone_number" json:"phone_number"`
 	ShippingAddress		string 		`form:"shipping_address" json:"shipping_address" sql:"type:VARCHAR CHARACTER SET utf8 COLLATE utf8_general_ci"`
-	Longitude			uint		`form:"longitude" json:"longitude"`
-	Latitude			uint		`form:"latitude" json:"latitude"`
+	Longitude			float32		`form:"longitude" json:"longitude"`
+	Latitude			float32		`form:"latitude" json:"latitude"`
 	ReceiverName		string		`form:"receiver_name" json:"receiver_name" sql:"type:VARCHAR CHARACTER SET utf8 COLLATE utf8_general_ci"`
 	UserId				uint		`json:"user_id"`
 }
